Refactor core engine into modules
